- renamed the toJSON operator
